+++ src/ext/ddtrace.h
@@ -9,7 +9,7 @@ char *request_init_hook;
 zend_bool ignore_missing_overridables;
 HashTable class_lookup;
 HashTable function_lookup;
-zend_bool debug_mode;
+zend_bool log_backtrace;
 user_opcode_handler_t ddtrace_old_fcall_handler;
 user_opcode_handler_t ddtrace_old_icall_handler;
 user_opcode_handler_t ddtrace_old_fcall_by_name_handler;