+++ CHANGELOG.md
@@ -9,8 +9,7 @@ All notable changes to this project will be documented in this file - [read more
 - Auto-instrumentation when user's autoloader throws exception on not found #305
 
 ### Added
-- Debug mode - enabled via ddtrace.debug_mode=1 ini setting (#300)
-- Print backtrace on segmentation fault in debug_mode (#300)
+- Log backtrace on segmentation fault, enabled via ddtrace.log_backtrace=1 (#300)
 
 ## [0.13.1]